requirements.txt
Flask==2.2.3
Jinja2==3.1.2
Werkzeug==2.2.2
python-dotenv
celery[redis]
gunicorn
cerberus